fixup rename module
