+++ apollo-router/src/plugins/connectors/tests/mod.rs
@@ -38,7 +38,7 @@ use crate::services::router::Request;
 use crate::services::supergraph;
 use crate::uplink::license_enforcement::LicenseState;
 
-mod batch;
+mod connect_on_type;
 mod mock_api;
 mod quickstart;
 #[allow(dead_code)]
